Load leader card backs

+++ zzImageConverter.php
@@ -1,9 +1,19 @@
 <?php
 
-function CheckImage($cardID, $url)
+function CheckImage($cardID, $url, $isBack=false)
 {
-  $filename = "./WebpImages/" . $cardID . ".webp";
-  $filenameNew = "./New Cards/" . $cardID . ".webp";
+  if($isBack) {
+    $filename = "./WebpImages/" . $cardID . "-Back.webp";
+    $filenameNew = "./New Cards/" . $cardID . "-Back.webp";
+    $concatFilename = "./concat/" . $cardID . "-Back.webp";
+    $cropFilename = "./crops/" . $cardID . "_cropped-Back.png";
+  }
+  else {
+    $filename = "./WebpImages/" . $cardID . ".webp";
+    $filenameNew = "./New Cards/" . $cardID . ".webp";
+    $concatFilename = "./concat/" . $cardID . ".webp";
+    $cropFilename = "./crops/" . $cardID . "_cropped.png";
+  }
   if(!file_exists($filename))
   {
     $imageURL = $url;
@@ -28,14 +38,13 @@ function CheckImage($cardID, $url)
       imagedestroy($image);
     }
   }
-  if(!file_exists($filenameNew)) {
+  if(!file_exists($filename) && !file_exists($filenameNew)) {
     echo("Converting image for " . $cardID . " to new format.<BR>");
     $image = imagecreatefromwebp($filename);
     //$image = imagecreatefrompng($filename);
     imagewebp($image, $filenameNew);
     imagedestroy($image);
   }
-  $concatFilename = "./concat/" . $cardID . ".webp";
   if(!file_exists($concatFilename))
   {
     echo("Concat image for " . $cardID . " does not exist. Converting: $filename<BR>");
@@ -60,7 +69,6 @@ function CheckImage($cardID, $url)
       if(file_exists($concatFilename)) echo("Image for " . $cardID . " successfully converted to concat.<BR>");
     }
   }
-  $cropFilename = "./crops/" . $cardID . "_cropped.png";
   if(!file_exists($cropFilename))
   {
     echo("Crop image for " . $cardID . " does not exist.<BR>");
